Feat: 이전글, 다음글 이동 구현

+++ src/main/java/com/ll/exam/article/repository/ArticleRepository.java
@@ -87,7 +87,7 @@ public class ArticleRepository {
         return sql.selectRow(ArticleDto.class);
     }
 
-    public ArticleDto getNextArticle(int id) {
+    public ArticleDto getNextArticle(long id) {
         SecSql sql = myMap.genSecSql();
         sql
                 .append("SELECT *")
